Adding two important variables.
On comma's I like have custom code, this so that I can process data as
we read in arguments. this is useful when we are reading from the
keyboard or file, and wont store the data into variable.
Current "input" implantation does work, on normal variables, but not on
arrays. This improvement should allow input to also work on arrays.